numpy/IMAGE_TOOL/requirements.txt
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
scipy>=1.7.0
